Move download_files.go to subcmd
